change amount

+++ api/v1/budgets/dtos/budget_dtos.go
@@ -25,7 +25,7 @@ type (
 		IDCategory               uuid.UUID `json:"category_id"`
 		IDSubCategory            uuid.UUID `json:"sub_category_id"`
 		IDMasterTransactionTypes uuid.UUID `json:"id_master_transaction_types,omitempty"`
-		Amount                   int       `json:"budget_amount"`
+		Amount                   int64     `json:"budget_amount"`
 		Departure                string    `json:"departure,omitempty"`
 		Arrival                  string    `json:"arrival,omitempty"`
 		TravelStartDate          string    `json:"travel_start_date,omitempty"`
